refactor: move middleware to api/, add compile-time interface checks
