+++ backend/api/rest/middleware.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/Asheze1127/progress-checker/backend/pkg/middleware"
+	"github.com/Asheze1127/progress-checker/backend/api/middleware"
 	"github.com/Asheze1127/progress-checker/backend/pkg/slack"
 )
